+++ src/config/config.go
@@ -8,6 +8,7 @@ import (
 	"github.com/dattito/purrmannplus-backend/utils"
 )
 
+// ENVIRONMENT VARIABLES
 var (
 	DOT_ENV_FILE_PATH                        string // Path to the .env file, only needed if USE_DOT_ENV_FILE is set to true
 	USE_DOT_ENV_FILE                         bool   // If true, the .env file will be loaded
@@ -19,6 +20,7 @@ var (
 	AUTHORIZATION_COOKIE_HTTPONLY            bool   // If true, the cookie will be set as httponly
 	AUTHORIZATION_COOKIE_SECURE              bool   // If true, the cookie will be set as secure
 	AUTHORIZATION_COOKIE_SAMESITE            string // Either "lax" (default), "strict", "disabled" or "none"
+	AUTHORIZATION_EXPIRATION_TIME            int    // The expiration time of the auth jwt tokens / auth cookies in seconds (default to 2678400 seconds = 1 month)
 	ENABLE_API                               bool   // If true, the api will be enabled, otherwise there will be no listener
 	ENABLE_SUBSTITUTIONS_SCHEDULER           bool   // If true, the substitutions scheduler will be enabled
 	SUBSTITUTIONS_UPDATECRON                 string // Cron expression for the substitutions scheduler
@@ -30,12 +32,15 @@ var (
 	SIGNAL_CLI_GRPC_API_URL                  string // The url of the signal cli grpc api
 	SIGNAL_SENDER_PHONENUMBER                string // The phonenumber of the signal sender
 	JWT_SECRET                               string // The secret used to sign the jwt tokens
-	DNT_JWT_RANDOM_SECRET                    string // Gets generated on init, so DO NOT TOUCH
 	SUBSTITUTION_URL                         string // The url of the substitution website
 	LOGGING_FILE                             string // The file to log to, if empty, logs to stdout
 	LOG_LEVEL                                int    // 0-5: 0:silent, 1:fatal, 2:error, 3:warn, 4:info, 5:debug
 	DNT_VERSION                              string // The version of this application. It is automatically set by docker, so DO NOT TOUCH
 )
+
+// END OF ENDVIRONMENT VARIABLES
+
+var JWT_SHORTLIVING_SECRET string
 
 func Init() error {
 
@@ -89,6 +94,11 @@ func Init() error {
 		return fmt.Errorf("AUTHORIZATION_COOKIE_SAMESITE must be one of lax, strict, disabled, none")
 	}
 
+	AUTHORIZATION_EXPIRATION_TIME, err = utils.GetIntEnv("AUTH_EXPIRATION_TIME", 2678400)
+	if err != nil {
+		return err
+	}
+
 	ENABLE_API, err = utils.GetBoolEnv("ENABLE_API", true)
 	if err != nil {
 		return err
@@ -127,7 +137,7 @@ func Init() error {
 		return err
 	}
 
-	DNT_JWT_RANDOM_SECRET = utils.GenerateString(128)
+	JWT_SHORTLIVING_SECRET = utils.GenerateString(128)
 
 	SUBSTITUTION_URL, err = utils.GetEnvInDev("SUBSTITUTION_URL", "https://vertretungsplan.hpg-speyer.de")
 	if err != nil {